rename files to match naming convention
